+++ lib/mainImpl.dart
@@ -65,6 +65,7 @@ import 'package:illinois/ui/widgets/FlexContent.dart';
 
 import 'package:rokwire_plugin/rokwire_plugin.dart';
 import 'package:rokwire_plugin/service/config.dart' as rokwire;
+import 'package:rokwire_plugin/service/content.dart';
 import 'package:rokwire_plugin/service/groups.dart';
 import 'package:rokwire_plugin/service/location_services.dart';
 import 'package:rokwire_plugin/service/app_navigation.dart';
@@ -130,6 +131,7 @@ void mainImpl({ rokwire.ConfigEnvironment? configEnvironment }) async {
     IlliniCash(),
     FlexUI(),
     Onboarding(),
+    Content(),
     Polls(),
     GeoFence(),
     Voter(),